refactor(control-api): extract handlers from response_actions/mod.rs
